+++ embassy-net-esp-hosted/src/lib.rs
@@ -16,9 +16,10 @@ use embassy_time::{Duration, Instant, Timer};
 use embedded_hal::digital::OutputPin;
 
 use crate::ioctl::{PendingIoctl, Shared};
+use crate::rpc::ioctl_ctx::IoctlMessage;
 use crate::rpc::{Backend, HostedEvent, RpcBackend};
 
-mod proto;
+pub(crate) mod proto;
 
 // must be first
 mod fmt;
@@ -119,14 +120,21 @@ enum InterfaceType {
 }
 
 const MAX_BUFFER_SIZE: usize = 1600;
-// Theoretical max overhead is 29 bytes. Biggest message currently is OTA write with 256 bytes.
-const MAX_IOCTL_SIZE: usize = 256 + 29;
+// Maximum payload size
+const MAX_IOCTL_SIZE: usize = if cfg!(feature = "esp-hosted-fg") {
+    // Scan results are unlimited
+    MAX_BUFFER_SIZE - 12
+} else {
+    // Theoretical max overhead is 29 bytes. Biggest message currently is OTA write with 256 bytes.
+    256 + 29
+};
 const HEARTBEAT_MAX_GAP: Duration = Duration::from_secs(20);
 
 /// State for the esp-hosted driver.
 pub struct State {
     shared: Shared,
     ioctl_buffer: [u8; MAX_IOCTL_SIZE],
+    msg_buffer: IoctlMessage,
     ch: ch::State<MTU, 4, 4>,
     #[cfg(feature = "bluetooth")]
     bt: bluetooth::BtState,
@@ -138,6 +146,7 @@ impl State {
         Self {
             shared: Shared::new(),
             ch: ch::State::new(),
+            msg_buffer: IoctlMessage::default(),
             ioctl_buffer: [0u8; MAX_IOCTL_SIZE],
             #[cfg(feature = "bluetooth")]
             bt: bluetooth::BtState::new(),
@@ -196,7 +205,7 @@ where
         net_device: device,
         #[cfg(feature = "bluetooth")]
         bluetooth: bt_driver,
-        control: Control::new(state_ch, &state.shared, &mut state.ioctl_buffer),
+        control: Control::new(state_ch, &state.shared, &mut state.ioctl_buffer, &mut state.msg_buffer),
         runner,
     }
 }